Move credential-specific auth code to individual classes
