+++ sdk/identity/identity/src/credentials/clientCertificateCredential.ts
@@ -1,32 +1,36 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+import qs from "qs";
+import jws from "jws";
+import uuid from "uuid";
 import { readFileSync } from "fs";
 import { createHash } from "crypto";
 import { TokenCredential, GetTokenOptions, AccessToken } from "@azure/core-http";
 import { IdentityClientOptions, IdentityClient } from "../client/identityClient";
 
+const SelfSignedJwtLifetimeMins = 10;
+
 /**
  * Enables authentication to Azure Active Directory using a PEM-encoded
  * certificate that is assigned to an App Registration.  More information
  * on how to configure certificate authentication can be found here:
- * 
+ *
  * https://docs.microsoft.com/en-us/azure/active-directory/develop/active-directory-certificate-credentials#register-your-certificate-with-azure-ad
  *
  */
 export class ClientCertificateCredential implements TokenCredential {
   private identityClient: IdentityClient;
-  private _tenantId: string;
-  private _clientId: string;
-  private _certificateString: string;
-
-  public certificateThumbprint: string;
-  public certificateX5t: string;
+  private tenantId: string;
+  private clientId: string;
+  private certificateString: string;
+  private certificateThumbprint: string;
+  private certificateX5t: string;
 
   /**
    * Creates an instance of the ClientCertificateCredential with the details
    * needed to authenticate against Azure Active Directory with a certificate.
-   * 
+   *
    * @param tenantId The Azure Active Directory tenant (directory) ID.
    * @param clientId The client (application) ID of an App Registration in the tenant.
    * @param certificatePath The path to a PEM-encoded public/private key certificate on the filesystem.
@@ -39,13 +43,13 @@ export class ClientCertificateCredential implements TokenCredential {
     options?: IdentityClientOptions
   ) {
     this.identityClient = new IdentityClient(options);
-    this._tenantId = tenantId;
-    this._clientId = clientId;
+    this.tenantId = tenantId;
+    this.clientId = clientId;
 
-    this._certificateString = readFileSync(certificatePath, "utf8");
+    this.certificateString = readFileSync(certificatePath, "utf8");
 
     const certificatePattern = /(-+BEGIN CERTIFICATE-+)(\n\r?|\r\n?)([A-Za-z0-9+/\n\r]+=*)(\n\r?|\r\n?)(-+END CERTIFICATE-+)/;
-    const matchCert = this._certificateString.match(certificatePattern);
+    const matchCert = this.certificateString.match(certificatePattern);
     const publicKey = matchCert ? matchCert[3] : "";
     if (!publicKey) {
       throw new Error(
@@ -66,7 +70,7 @@ export class ClientCertificateCredential implements TokenCredential {
    * successful.  If authentication cannot be performed at this time, this method may
    * return null.  If an error occurs during authentication, an {@link AuthenticationError}
    * containing failure details will be thrown.
-   * 
+   *
    * @param scopes The list of scopes for which the token will have access.
    * @param options The options used to configure any requests this
    *                TokenCredential implementation might make.
@@ -75,13 +79,58 @@ export class ClientCertificateCredential implements TokenCredential {
     scopes: string | string[],
     options?: GetTokenOptions
   ): Promise<AccessToken | null> {
-    return this.identityClient.authenticateClientCertificate(
-      this._tenantId,
-      this._clientId,
-      this._certificateString,
-      this.certificateX5t,
-      scopes,
-      options
-    );
+    const tokenId = uuid.v4();
+    const audienceUrl = `${this.identityClient.authorityHost}/${this.tenantId}/oauth2/v2.0/token`;
+    const header: jws.Header = {
+      typ: "JWT",
+      alg: "RS256",
+      x5t: this.certificateX5t
+    };
+
+    const payload = {
+      iss: this.clientId,
+      sub: this.clientId,
+      aud: audienceUrl,
+      jti: tokenId,
+      nbf: this.timestampInSeconds(new Date()),
+      exp: this.timestampInSeconds(this.addMinutes(new Date(), SelfSignedJwtLifetimeMins))
+    };
+
+    const clientAssertion = jws.sign({
+      header,
+      payload,
+      secret: this.certificateString
+    });
+
+    const webResource = this.identityClient.createWebResource({
+      url: audienceUrl,
+      method: "POST",
+      disableJsonStringifyOnBody: true,
+      deserializationMapper: undefined,
+      body: qs.stringify({
+        response_type: "token",
+        grant_type: "client_credentials",
+        client_id: this.clientId,
+        client_assertion_type: "urn:ietf:params:oauth:client-assertion-type:jwt-bearer",
+        client_assertion: clientAssertion,
+        scope: typeof scopes === "string" ? scopes : scopes.join(" ")
+      }),
+      headers: {
+        Accept: "application/json",
+        "Content-Type": "application/x-www-form-urlencoded"
+      },
+      abortSignal: options && options.abortSignal
+    });
+
+    return this.identityClient.sendTokenRequest(webResource);
+  }
+
+  private timestampInSeconds(date: Date): number {
+    return Math.floor(date.getTime() / 1000);
+  }
+
+  private addMinutes(date: Date, minutes: number): Date {
+    date.setMinutes(date.getMinutes() + minutes);
+    return date;
   }
 }
