+++ sdk/identity/identity/src/credentials/clientSecretCredential.ts
@@ -1,6 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+import qs from "qs";
 import { TokenCredential, GetTokenOptions, AccessToken } from "@azure/core-http";
 import { IdentityClientOptions, IdentityClient } from "../client/identityClient";
 
@@ -8,21 +9,21 @@ import { IdentityClientOptions, IdentityClient } from "../client/identityClient"
  * Enables authentication to Azure Active Directory using a client secret
  * that was generated for an App Registration.  More information on how
  * to configure a client secret can be found here:
- * 
+ *
  * https://docs.microsoft.com/en-us/azure/active-directory/develop/quickstart-configure-app-access-web-apis#add-credentials-to-your-web-application
  *
  */
 export class ClientSecretCredential implements TokenCredential {
   private identityClient: IdentityClient;
-  private _tenantId: string;
-  private _clientId: string;
-  private _clientSecret: string;
+  private tenantId: string;
+  private clientId: string;
+  private clientSecret: string;
 
   /**
    * Creates an instance of the ClientSecretCredential with the details
    * needed to authenticate against Azure Active Directory with a client
    * secret.
-   * 
+   *
    * @param tenantId The Azure Active Directory tenant (directory) ID.
    * @param clientId The client (application) ID of an App Registration in the tenant.
    * @param clientSecret A client secret that was generated for the App Registration.
@@ -35,9 +36,9 @@ export class ClientSecretCredential implements TokenCredential {
     options?: IdentityClientOptions
   ) {
     this.identityClient = new IdentityClient(options);
-    this._tenantId = tenantId;
-    this._clientId = clientId;
-    this._clientSecret = clientSecret;
+    this.tenantId = tenantId;
+    this.clientId = clientId;
+    this.clientSecret = clientSecret;
   }
 
   /**
@@ -45,7 +46,7 @@ export class ClientSecretCredential implements TokenCredential {
    * successful.  If authentication cannot be performed at this time, this method may
    * return null.  If an error occurs during authentication, an {@link AuthenticationError}
    * containing failure details will be thrown.
-   * 
+   *
    * @param scopes The list of scopes for which the token will have access.
    * @param options The options used to configure any requests this
    *                TokenCredential implementation might make.
@@ -54,12 +55,25 @@ export class ClientSecretCredential implements TokenCredential {
     scopes: string | string[],
     options?: GetTokenOptions
   ): Promise<AccessToken | null> {
-    return this.identityClient.authenticateClientSecret(
-      this._tenantId,
-      this._clientId,
-      this._clientSecret,
-      scopes,
-      options
-    );
+    const webResource = this.identityClient.createWebResource({
+      url: `${this.identityClient.authorityHost}/${this.tenantId}/oauth2/v2.0/token`,
+      method: "POST",
+      disableJsonStringifyOnBody: true,
+      deserializationMapper: undefined,
+      body: qs.stringify({
+        response_type: "token",
+        grant_type: "client_credentials",
+        client_id: this.clientId,
+        client_secret: this.clientSecret,
+        scope: typeof scopes === "string" ? scopes : scopes.join(" ")
+      }),
+      headers: {
+        Accept: "application/json",
+        "Content-Type": "application/x-www-form-urlencoded"
+      },
+      abortSignal: options && options.abortSignal
+    });
+
+    return this.identityClient.sendTokenRequest(webResource);
   }
 }
