+++ sdk/identity/identity/test/credentials/managedIdentityCredential.spec.ts
@@ -4,7 +4,7 @@
 import qs from "qs";
 import assert from "assert";
 import { ManagedIdentityCredential } from "../../src";
-import { ImdsEndpoint, ImdsApiVersion, AppServiceMsiApiVersion } from "../../src/client/identityClient";
+import { ImdsEndpoint, ImdsApiVersion, AppServiceMsiApiVersion } from "../../src/credentials/managedIdentityCredential";
 import { MockAuthHttpClient, MockAuthHttpClientOptions } from "./authTestUtils";
 import { WebResource, AccessToken } from "@azure/core-http";
